+++ client_web/package.json
@@ -14,6 +14,7 @@
     "@fluentui/react-icons": "^2.0.239",
     "@monaco-editor/react": "^4.6.0",
     "@swc/core": "^1.5.5",
+    "katex": "^0.16.10",
     "monaco-editor": "^0.48.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
