+++ client_web/src/router/Root.tsx
@@ -10,19 +10,21 @@ import NavBar from "./Navbar";
 import "../css/style.css";
 import useWebSocket from "react-use-websocket";
 import * as globals from "./Globals"
+import { useCallback, useEffect } from "react";
 
 export default function Root() {
     const {
         sendJsonMessage,
         lastJsonMessage,
         readyState
-    } = useWebSocket("ws://" + location.host + "/wsapi", {
+    } = useWebSocket("https://" + location.host + "/wss", {
         share: true,
         shouldReconnect: () => true,
         reconnectAttempts: 10,
         reconnectInterval: 3000,
         onClose: () => {
-            if (globals.fetchData("isLoggedIn")) {
+            if (globals.fetchData("isLoggedIn") === true) {
+                console.log("quitting...");
                 sendJsonMessage({
                     type: "quit",
                     content: {
@@ -35,6 +37,27 @@ export default function Root() {
             }
         },
     });
+
+    const beforeunload = useCallback((ev: Event) => {
+        if (ev) {
+            if (globals.fetchData("isLoggedIn") === true) {
+                console.log("quitting...");
+                sendJsonMessage({
+                    type: "quit",
+                    content: {
+                        username: globals.fetchData("loginUsername"),
+                        session_token: globals.fetchData("sessionToken"),
+                        request_key: globals.randomUUID()
+                    }
+                });
+                globals.setData("isLoggedIn", false);
+            }
+        }
+    }, [sendJsonMessage]);
+
+    useEffect(() => {
+        window.addEventListener('beforeunload', beforeunload);
+    }, [beforeunload])
 
     return (
         <FluentProvider theme={webLightTheme}>
