+++ problem/1001/output/README.md
@@ -1,0 +1,3 @@
+# README
+
+**This is a directory only for temporarily storation. DO NOT put ANYTHING important here.**
